Aswathy/ Refactored the script for CIO (#2856)
* refactor: refactored the script for CIO

* refactor: added the addscript in utility

* refactor: removed the code smell issue

* refactor: refactoring script

* refactor: reused the addscript from utility.js

* fix: refresh checkbox issue for submit status

* fix: code smell issue

* refactor: changes in the addscript utility
