+++ src/pages/academy/components/_side-subscription-banner.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react'
 import styled from 'styled-components'
 import { TextWrapper } from './_common'
+import { addScriptForCIO } from './utility'
 import AgreementLabel from 'components/custom/_agreement-label'
 import validation from 'common/validation'
 import { localize, Localize } from 'components/localization'
@@ -11,6 +12,7 @@ import device from 'themes/device'
 import { DerivStore } from 'store'
 import EmailSubscriptionSVG from 'images/svg/blog/articles/blog-article-email-subscription.svg'
 import CrossIcon from 'images/svg/help/cross.svg'
+import { getCountryRule } from 'components/containers/visibility'
 
 const SubscribeBannerWrapper = styled(Flex)`
     max-width: 282px;
@@ -125,12 +127,12 @@ const SideSubscriptionBanner = () => {
     const [email_error_msg, setEmailErrorMsg] = React.useState('')
     const [name_error_msg, setNameErrorMsg] = React.useState('')
     const [submit_error_msg, setSubmitErrorMsg] = React.useState('')
-
-    const { is_eu_country, user_country } = React.useContext(DerivStore)
+    const { user_country } = React.useContext(DerivStore)
+    const { is_eu } = getCountryRule()
 
     useEffect(() => {
         if (!window._cio) {
-            addScriptForCIO()
+            addScriptForCIO(is_eu)
         }
         const url = 'https://assets.customer.io/assets/track.js'
         fetch(url, {
@@ -144,41 +146,6 @@ const SideSubscriptionBanner = () => {
                 setSubmitStatus(false)
             })
     }, [])
-
-    const addScriptForCIO = () => {
-        const addScript = (settings) => {
-            const script = document.createElement('script')
-            const { async, text, src, id } = settings
-
-            if (async) script.async = settings['async']
-            if (text) script.text = settings['text']
-            if (src) script.src = settings['src']
-            if (id) script.id = settings['id']
-            document.body.appendChild(script)
-        }
-        const site_id = process.env.GATSBY_ENV_CIO_SITE_ID
-
-        let cio_url = 'https://assets.customer.io/assets/track.js'
-        if (is_eu_country) {
-            cio_url = 'https://assets.customer.io/assets/track-eu.js'
-        }
-
-        addScript({
-            text: `
-            var _cio = _cio || [];
-            var a,b,c;a=function(f){return function(){_cio.push([f].
-            concat(Array.prototype.slice.call(arguments,0)))}};b=["load","identify",
-            "sidentify","track","page"];for(c=0;c<b.length;c++){_cio[b[c]]=a(b[c])};
-            var t = document.createElement('script'),
-                s = document.getElementsByTagName('script')[0];
-            t.async = true;
-            t.id    = 'cio-tracker';
-            t.setAttribute('data-site-id', '${site_id}');
-            t.src = '${cio_url}'
-            //If your account is in the EU, use:
-            s.parentNode.insertBefore(t, s);`,
-        })
-    }
 
     const handleChange = (event) => {
         setChecked(event.currentTarget.checked)
